Update documentation and add new utilities
Documentation for `README.md` files was updated to reflect changes in the API for image transformations and the renaming of utility classes. A new utility function `funcDefined` was added to `lib.js` for checking function definition. Furthermore, comments have been added for better understanding of the `DataManagerReadModel` in the `Hipot` model classes.