Change 'INCLUDE_SUBSECTIONS' value in filter configuration
The filter configuration in the 'hipot/iblock.section' class has been modified. Previously, it included sub-sections when determining the count of elements in a section. Now, it does not include sub-sections, as demonstrated by 'INCLUDE_SUBSECTIONS' now being set to 'N' instead of 'Y'. This change applies to both Hipot projects.