Update caching implementation and optimize code
This commit  broadens the type of acceptable cache engines to cache engine interfaces and presents an exception when unable to connect to Memcache. This also includes the necessary changes in the related documentation.

it updates the caching system by disabling the automatic start of PhpCacher's tagged cache and removing the default cache service for PHPCacher. It also optimizes a condition check in the event handler and tweaks the style settings in the UI. Addtionally,