Add new SocialCard-dark.html file
The new file, SocialCard-dark.html, has been added to the docs directory. This file contains the HTML, CSS and JavaScript structure to present a contributor's GitHub stats in a dark-theme social card format. It is meant to provide a stylized and interactive way for contributors to display their GitHub statistics.