Add function to get payment system service by order
The function 'getPaySystemServiceByOrder' has been added to the 'Sale' class. This function retrieves the payment system service for a given order, returning these as an array. The necessary exceptions for possible errors have been provided as well. Further error handling may be necessary depending on the context in which this function is used.