Now IblockAbstractLayer has numeric property ID in phpDocs.
introduced a new method in EventHelper.php for setting property values efficiently.

Added `secondsToTimeString` and `timeToIso8601Duration` functions in DateTimeUtils.php for better time conversion (it's need to rewrite, todos)

Add methods to clean unused basket properties on events

Integrated two event handlers in `handlers_add.php` to trigger the cleanup of unused basket properties upon adding an item to the basket and upon saving an order. Also, implemented the `deleteUnUsedBasketProps` method in `Sale.php` to execute the actual deletion in the database.